optimized PKExistingEdgePartitionBuilder

+++ src/main/scala/org/apache/spark/graphx/pkgraph/graph/PKEdge.scala
@@ -2,7 +2,7 @@ package org.apache.spark.graphx.pkgraph.graph
 
 case class PKEdge[E](
     var index: Int = -1,
-    var line: Long = 0,
-    var col: Long = 0,
+    var line: Int = 0,
+    var col: Int = 0,
     var attr: E = null.asInstanceOf[E]
 )
